test: remove legacy pubmed tests
