+++ tests/test_openalex_crossref_library.py
@@ -16,7 +16,7 @@ def test_fetch_openalex_uses_cfg(monkeypatch) -> None:
         called["timeout"] = timeout
         return {}, ""
 
-    monkeypatch.setattr("library.pubmed_library._do_request", fake_do_request)
+    monkeypatch.setattr("library.pubmed.query._do_request", fake_do_request)
 
     cfg = Config(
         api=ApiCfg(user_agent="test@example.com"),
@@ -46,7 +46,7 @@ def test_fetch_crossref_uses_cfg(monkeypatch) -> None:
         called["timeout"] = timeout
         return {}, ""
 
-    monkeypatch.setattr("library.pubmed_library._do_request", fake_do_request)
+    monkeypatch.setattr("library.pubmed.query._do_request", fake_do_request)
 
     cfg = Config(
         api=ApiCfg(user_agent="test@example.com"),
@@ -74,7 +74,7 @@ def test_rate_limiter_shared(monkeypatch) -> None:
         delays.append(delay)
 
     monkeypatch.setattr(rl, "sleep", fake_sleep)
-    monkeypatch.setattr("library.pubmed_library._do_request", lambda *a, **k: ({}, ""))
+    monkeypatch.setattr("library.pubmed.query._do_request", lambda *a, **k: ({}, ""))
 
     cfg = Config(
         api=ApiCfg(user_agent="test@example.com"),
